Fix array length mismatch in tb_075
The assignment `ac <= a_t(c)` fails because the lengths of `y` field are
different.  The comment says this is invalid, but I'm not sure how or if
we can express an expected-to-fail test in VUnit so I just commented
this line out.

Also improved the test a bit so it checks the result of the valid
conversion.